Fixed a bug where a name containing unicode chars would not be recognized as already being in the database.
Explanation : the name in the database is available as an unicode
object, but ftplib return str objects. If they contains accents,
they are not considered equals.

I manually did the conversion where it was needed.

Unicode can be a bitch in python 2, because it looks like it works
until it doesn't.